03_backend_web_development_and_api/07_mvc/1e9d98ad0071f188d1b8ba88647e6c1ce8f243e5/02_warmup_classy_cat/src/index.ts
export class Cat {
  name: string;
  color: string;
  age: number;

  constructor(name: string, color: string, age: number) {
    this.name = name;
    this.color = color;
    this.age = age;
  }

  purr() {
    return `${this.name}, the cat, purrs...`;
  }

  meow() {
    return `${this.name}, the ${this.color} cat, meows!`;
  }

  attack(withSuccess: boolean) {
    if (withSuccess) {
      console.log("You've been scratched!");
    } else {
      console.log(`Wow, ${this.name} missed you, how lucky!`);
    }
  }
}
